+++ panelitems/engine_rpm.cpp
@@ -75,8 +75,8 @@ void EngineRPM::loadSettings(QSettings &settings) {
     QString unitname = settings.value("unit").toString();
     VelocityUnit unit = Units::velocityUnitForName(unitname);
     setUnit(unit);
-    setMaxValue(settings.value("maxvalue", 300).toFloat());
-    setNumberScale(settings.value("scalevalue", 1.).toFloat());
+    setMaxValue(settings.value("maxvalue", 300).toDouble());
+    setNumberScale(settings.value("scalevalue", 1.).toDouble());
 }
 
 void EngineRPM::setMaxValue(float mv) {
