+++ panelitems/variometer.cpp
@@ -48,7 +48,7 @@ void Variometer::loadSettings(QSettings &settings) {
     QString unitname = settings.value("unit").toString();
     VelocityUnit unit = Units::velocityUnitForName(unitname);
     setUnit(unit);
-    setMaxValue(settings.value("maxvalue", 300).toFloat());
+    setMaxValue(settings.value("maxvalue", 300).toDouble());
     setIsTotalEnergy(settings.value("totalenergy", false).toBool());
 }
